+++ cocos/editor-support/cocostudio/ActionTimeline/CCActionTimeline.cpp
@@ -199,7 +199,7 @@ void ActionTimeline::step(float delta)
         return;
     }
     _time += delta * _timeSpeed;
-    float deltaCurrFrameTime = std::abs(_time - _currentFrame * _frameInternal);
+    float deltaCurrFrameTime = (float)std::abs(_time - _currentFrame * _frameInternal);
     if (deltaCurrFrameTime < _frameInternal)
         return;
