+++ cocos/editor-support/cocostudio/CCDatas.cpp
@@ -124,8 +124,8 @@ void BaseData::subtract(BaseData *from, BaseData *to, bool limit)
 
     if (to->tweenRotate)
     {
-        skewX += to->tweenRotate * M_PI * 2;
-        skewY -= to->tweenRotate * M_PI * 2;
+        skewX += to->tweenRotate * (float)M_PI * 2;
+        skewY -= to->tweenRotate * (float)M_PI * 2;
     }
 	
 }
